requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gspread==6.0.2
google-auth==2.27.0
pydantic==2.6.1
httpx==0.27.0
